refactor: auth add: jwt dependencies, database config, spring security

+++ library-system-backend/src/main/java/com/rige/services/impl/UserServiceImpl.java
@@ -1,41 +1,41 @@
 package com.rige.services.impl;
 
+import com.rige.dto.response.UserResponse;
 import com.rige.entities.UserEntity;
+import com.rige.mappers.UserMapper;
 import com.rige.repositories.UserRepository;
 import com.rige.services.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
-import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final UserMapper userMapper;
 
     @Override
-    public List<UserEntity> findAllUsers() {
-        return userRepository.findAll();
+    public List<UserResponse> findAllUsers() {
+        return userRepository.findAll().stream()
+                .map(userMapper::toResponse)
+                .collect(Collectors.toList());
     }
 
     @Override
-    public Optional<UserEntity> findUserById(Long id) {
-        return userRepository.findById(id);
-    }
-
-    @Override
-    public UserEntity findUserByEmail(String email) {
-        return userRepository.findByEmail(email);
-    }
-
-    @Override
-    public UserEntity saveUser(UserEntity user) {
-        return userRepository.save(user);
+    public UserResponse findUserById(Long id) {
+        UserEntity user = userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+        return userMapper.toResponse(user);
     }
 
     @Override
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("User not found.");
+        }
         userRepository.deleteById(id);
     }
 }
